feat: set default hasher to hashtree for benchmark
