+++ setHasher.mjs
@@ -1,5 +1,5 @@
-// Set the hasher to as-sha256
-// Used to run benchmarks with with visibility into as-sha256 performance, useful for Lodestar
+// Set the hasher to hashtree
+// Used to run benchmarks with with visibility into hashtree performance, useful for Lodestar
 import {setHasher} from "@chainsafe/persistent-merkle-tree/lib/hasher/index.js";
-import {hasher} from "@chainsafe/persistent-merkle-tree/lib/hasher/as-sha256.js";
+import {hasher} from "@chainsafe/persistent-merkle-tree/lib/hasher/hashtree.js";
 setHasher(hasher);
